Remove duplicated V2 test suites.
